+++ src/mods/data.lua
@@ -1,7 +1,0 @@
-local data = {}
-
-function data.buildStorage()
-    return {}
-end
-
-return data